moved get creds to middleware, added auth middleware (checks session)
