--- web_server/handlers/adapters.go
@@ -1,1 +1,0 @@
-package handlers
